requirements.txt
# Core ML Libraries
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API & Data Collection
requests>=2.31.0
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Logging & Testing
structlog>=23.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0